fix(blackjack): 플레이어 출력 순서 보장, 상금 정수형으로 변경

+++ src/main/java/blackjack/domain/BettingBox.java
@@ -15,7 +15,7 @@ public class BettingBox {
         bettingBox.put(guest, bettingMoney);
     }
 
-    public double getPrizeMoney(Player guest, WinDrawLose winDrawLose, boolean isBlackjack) {
+    public int getPrizeMoney(Player guest, WinDrawLose winDrawLose, boolean isBlackjack) {
         BettingMoney bettingMoney = bettingBox.get(guest);
         if (winDrawLose == WinDrawLose.DRAW) {
             return 0;
@@ -26,9 +26,9 @@ public class BettingBox {
         return bettingMoney.getAmount() * -1;
     }
 
-    private double getWinningGuestPrizeMoney(BettingMoney bettingMoney, boolean isBlackjack) {
+    private int getWinningGuestPrizeMoney(BettingMoney bettingMoney, boolean isBlackjack) {
         if (isBlackjack) {
-            return bettingMoney.getAmount() * 1.5;
+            return (int)(bettingMoney.getAmount() * 1.5);
         }
         return bettingMoney.getAmount();
     }
